assembly: add assemblyutils
The implementation of TRACE2 tracing will require use of the
TryGetAssemblyVersion method. To prepare for this, move this method out
of the DiagnoseCommand class and into its own static class.
